resolve config when merge media  into more like this media
