+++ app/src/main/java/com/berlin/aflami/di/dataSourceModule.kt
@@ -4,11 +4,11 @@ import com.berlin.local.dao.SearchDao
 import com.berlin.local.datasource.SearchLocalDataSourceImpl
 import com.berlin.remote.MovieDetailsRemoteDataSourceImpl
 import com.berlin.remote.SearchRemoteDataSourceImpl
-import com.berlin.remote.SeriesDetailsRemoteDataSourceImpl
+import com.berlin.remote.TvShowDetailsRemoteDataSourceImpl
 import com.berlin.repository.datasource.local.SearchLocalDataSource
 import com.berlin.repository.datasource.remote.MovieDetailsRemoteDataSource
 import com.berlin.repository.datasource.remote.SearchRemoteDataSource
-import com.berlin.repository.datasource.remote.SeriesDetailsRemoteDataSource
+import com.berlin.repository.datasource.remote.TvShowDetailsRemoteDataSource
 import io.ktor.client.HttpClient
 import org.koin.dsl.module
