--- data/repository/src/main/java/com/berlin/repository/datasource/remote/SeriesDetailsRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package com.berlin.repository.datasource.remote
-
-import com.berlin.repository.datasource.remote.dto.TVShowResponse
-
-interface SeriesDetailsRemoteDataSource {
-    suspend fun getSeriesSimilar(seriesId: Long): TVShowResponse
-}
