--- domain/usecase/src/main/kotlin/repository/SeriesDetailsRepository.kt
@@ -1,7 +1,0 @@
-package repository
-
-import com.berlin.entity.TVShow
-
-interface SeriesDetailsRepository {
-    suspend fun getSeriesSimilar(seriesId:Long):List<TVShow>
-}
