+++ domain/usecase/src/main/kotlin/repository/TvShowDetailsRepository.kt
@@ -1,11 +1,13 @@
 package repository
 
 import com.berlin.entity.MediaCast
+import com.berlin.entity.TVShow
 import com.berlin.entity.TvShowDetails
 
 interface TvShowDetailsRepository {
     suspend fun getTvShowDetails(id: Long, language: String): TvShowDetails?
     suspend fun getSeriesImages(id:Long):List<String>
     suspend fun getSeriesCastDetails(seriesId: Long, language: String): List<MediaCast>
+    suspend fun getSeriesSimilar(seriesId:Long):List<TVShow>
 
 }
